Add one more switch, eeprom settings saving on powerloss and one more menu

+++ src/timer.h
@@ -1,11 +1,13 @@
+// General purpose timer to count time since the device power-up
+
 #pragma once
 
 #include <stdint.h>
 
-#define TIMER_TICKS_PER_SEC ( F_CPU / 256 )
+#define TIMER_TICKS_PER_SEC ( F_CPU / 256 ) // timer precision is 16 us
 
 void timer_start( void );
 void timer_stop( void );
 
-uint64_t timer_get_us( void );
-uint64_t timer_get_ms( void );
+uint32_t timer_get_us( void );
+uint32_t timer_get_ms( void );
